Adding a first option to scale.
